+++ R/count.R
@@ -84,7 +84,7 @@ NULL
 #' 
 #' debates <- partition_bundle(
 #'   "GERMAPARLMINI", s_attribute = "date", values = NULL,
-#'   regex = TRUE, mc = FALSE, verbose = FALSE
+#'   mc = FALSE, verbose = FALSE
 #' )
 #' y <- count(debates, query = "Arbeit", p_attribute = "word")
 #' y <- count(debates, query = c("Arbeit", "Migration", "Zukunft"), p_attribute = "word")