+++ NEWS.md
@@ -71,9 +71,8 @@ is offered (#110).
 - The `corpus` class includes a (new) slot `size`, just as the `regions` and the `subcorpus` classes.
 - The `split()`-method for `corpus` objects now accepts the argument `xml`, to indicate whether the annotation structure of the corpus is flat or nested.
 - The definition of the S4 class `partition` now includes a prototype defining default values for the slots 'stat' (a `data.table`) and the slot 'size' (`NA_integer_`). This avoids that an incomplete initialization of a `partition` object will result in an error.
-- Internally, the `partition_bundle()`-methods defined for `character`, `corpus` and `partition` objects now
-call the `split()`-methods for `corpus` und `subcorpus` objects, resulting in a huge performance gain (#112).
-
+- Internally, the `partition_bundle()`-methods defined for `character`, `corpus` and `partition` objects now call the `split()`-methods for `corpus` und `subcorpus` objects, resulting in a huge performance gain (#112).
+- The `merge()`-method is implemented for `subcorpus_bundle` objects now, and has been implemented for `subcorpus` objects.
 
 polmineR 0.8.0
 ==============